requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
joblib

# Backend
fastapi
uvicorn

# Face Recognition
opencv-python
torch
torchvision
facenet-pytorch
Pillow

# Optional (if using file upload)
python-multipart